Dimitar Zhekov <jimmy@is-vn.bg>: fixed the STONE install module to only offer filesystem creation if the user-space tools are present
git-svn-id: https://svn.exactcode.de/t2/trunk@1714 c5f82cb5-29bc-0310-9cd0-bff59a50e3bc

+++ package/base/sysfiles/stone_mod_install.sh
@@ -58,11 +58,15 @@ part_mkfs() {
 	cmd="$cmd 'reiserfs (journaling filesystem)'"
 	cmd="$cmd 'mkreiserfs /dev/$1/$2'"
 
-	cmd="$cmd 'IBM JFS  (journaling filesystem)'"
-	cmd="$cmd 'jfs_mkfs /dev/$1/$2'"
+	if type -p jfs_mkfs > /dev/null ; then
+		cmd="$cmd 'IBM JFS  (journaling filesystem)'"
+		cmd="$cmd 'jfs_mkfs /dev/$1/$2'"
+	fi
 
-	cmd="$cmd 'SGI XFS  (journaling filesystem)'"
-	cmd="$cmd 'mkfs.xfs /dev/$1/$2'"
+	if type -p mkfs.xfs > /dev/null ; then
+		cmd="$cmd 'SGI XFS  (journaling filesystem)'"
+		cmd="$cmd 'mkfs.xfs /dev/$1/$2'"
+	fi
 
 	eval "$cmd" && part_mount $1 $2
 }
